+++ src/main/java/com/newput/testCase/ServiceTestCase.java
@@ -1,12 +1,19 @@
 package com.newput.testCase;
 
+import static org.junit.Assert.assertEquals;
+
+import java.io.File;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.List;
+
 import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-
 
 import com.newput.domain.DateSheet;
 import com.newput.domain.Employee;
@@ -18,8 +25,6 @@ import com.newput.mapper.DateSheetMapper;
 import com.newput.mapper.EmployeeMapper;
 import com.newput.mapper.SessionMapper;
 import com.newput.service.EmpService;
-import com.newput.service.LoginService;
-import com.newput.service.SelectiveExcel;
 import com.newput.service.TSchedualService;
 import com.newput.utility.EMailSender;
 import com.newput.utility.ExcelTimeSheet;
@@ -27,18 +32,9 @@ import com.newput.utility.JsonResService;
 import com.newput.utility.ReqParseService;
 import com.newput.utility.TTUtil;
 
-import static org.junit.Assert.*;
-import java.io.File;
-import java.math.BigDecimal;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
-
-import java.util.List;
-
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(locations = "/applicationContext.xml")
-public class UnitTestCase {
+public class ServiceTestCase {
 
 	@Autowired
 	private DateSheet dateSheet;
@@ -54,7 +50,7 @@ public class UnitTestCase {
 
 	@Autowired
 	private Employee emp;
-	//
+	
 	@Autowired
 	private EmpService empService;
 	
@@ -66,9 +62,6 @@ public class UnitTestCase {
 
 	@Autowired
 	private JsonResService jsonResService;
-	//
-	@Autowired
-	private LoginService loginService;
 	
 	@Autowired
 	private TTUtil util;
@@ -80,9 +73,6 @@ public class UnitTestCase {
 	private ExcelTimeSheet excelTimeSheet;
 	
 	@Autowired
-	private SelectiveExcel excel;
-	
-	@Autowired
 	EmployeeMapper empMapper;
 
 	@Autowired
@@ -92,9 +82,7 @@ public class UnitTestCase {
 	public Long getCurrentTime() {
 		return System.currentTimeMillis() / 1000;
 	}
-
-	//String email = "modideepti37@gmail.com";
-	//String password = "abcd"; 
+	
 	String email = "rahul@newput.com";
 	String password = "abcd";
 	String empId = "1";
@@ -107,116 +95,6 @@ public class UnitTestCase {
 	String gender = "f";
 	String month = "October";
 	String year = "2015";
-	
-	
-	@Test
-	@Ignore
-	public void testRegisterUser(){
-		try{
-			SimpleDateFormat sdf = new SimpleDateFormat("dd-MM-yyyy");
-			emp.setFirstName(firstName);
-			emp.setLastName(lastName);
-			emp.setEmail(email);		
-			Date userDob = sdf.parse(dob);
-			Date userDoj = sdf.parse(doj);		
-			emp.setDob(userDob);
-			emp.setDoj(userDoj);
-			emp.setAddress(address);
-			emp.setContact(contact);
-			emp.setGender(gender);
-			String getPassword = util.md5(password);
-			emp.setPassword(getPassword);
-			emp.setStatus(false);
-			emp.setPasswordVerification(false);
-			emp.setRole("guest");
-			emp.setCreated(getCurrentTime());
-			emp.setTimeZone(new BigDecimal("5.5"));
-			String token = util.generateRandomString();		
-			emp.setvToken(token);			
-		}catch(Exception e){}
-		empService.addUser(emp);
-		assertEquals(true, jsonResService.isSuccess());
-		if (jsonResService.isSuccess()) {
-			String sendMail = emailSend.sendMail("registration");
-			assertEquals(null, sendMail);
-		}
-	}
-	
-	@Test
-	@Ignore
-	public void testMailVerification(){
-		emp.setvToken("3686");
-		emp.setEmail(email);
-		empService.mailVerify(emp);
-		assertEquals(true, jsonResService.isSuccess());
-	}
-	
-	@Test	
-	//@Ignore
-	public void testLogin() { 
-		emp.setEmail(email);
-		emp.setPassword(password);
-		loginService.createSession(emp);
-		assertEquals(true, jsonResService.isSuccess());
-	}	
-
-	@Test
-	@Ignore
-	public void testTimeEntry() throws ParseException {
-		timeSchedual.timeSheetValue("12:00", "9:00", "19:30", "06-10-2015", "12:30", "21:00", "23:00", Integer.parseInt(empId));
-		reqParser.setDateSheetValue("this is my 6 date", "06-10-2015", Integer.parseInt(empId));
-		timeSchedual.dateSheetValue();
-		assertEquals(true, jsonResService.isSuccess());
-	}
-	
-	@Test
-	@Ignore
-	public void testForgotPwd(){
-		emp.setEmail(email);
-		String ptoken = util.generateRandomString();
-		empService.resetPassword(email, ptoken, "password");
-		assertEquals(true, jsonResService.isSuccess());
-		if (jsonResService.isSuccess()) {
-			String sendMail = emailSend.sendMail("password");
-			assertEquals(null, sendMail);
-		}
-	}
-	
-	@Test
-	@Ignore
-	public void testExcelExport(){
-		assertEquals(true, util.validCheck(month, year));
-		File file = excelTimeSheet.createExcelSheet(Integer.parseInt(empId), month, year);
-		assertEquals(true, jsonResService.isSuccess());
-		file.delete();
-	}
-	
-	@Test
-	@Ignore
-	public void testPasswordVerification(){
-		emp.setId(Integer.parseInt(empId));
-		emp.setpToken("4669");
-		String newPassword = util.md5("rahul");
-		emp.setPassword(newPassword);
-		emp.setUpdated(getCurrentTime());
-		empService.pwdVerify(emp);
-		assertEquals(true, jsonResService.isSuccess());
-	}
-	
-	@Test
-	@Ignore
-	public void testMonthlyExcel(){
-		assertEquals(true, util.validCheck(month, year));
-		excel.monthSheet(month, Integer.parseInt(empId), year);
-		assertEquals(true, jsonResService.isSuccess());
-		assertNotNull(jsonResService.getData());		
-	}
-	
-	
-	/*
-	 * Test Cases for SERVICE CLASS
-	 */
-	
 	
 	@Test
 	@Ignore
@@ -371,5 +249,5 @@ public class UnitTestCase {
 			String sendMail = emailSend.sendMail("password");
 			assertEquals(null, sendMail);
 		}		
-	}		
+	}
 }
